% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_callback_step.R
\name{CallbackOutput}
\alias{CallbackOutput}
\title{Workflow CallbackOutput class}
\description{
Output for a callback step.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{output_name}}{The output name}

\item{\code{output_type}}{The output type}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CallbackOutput$new()}}
\item \href{#method-to_request}{\code{CallbackOutput$to_request()}}
\item \href{#method-expr}{\code{CallbackOutput$expr()}}
\item \href{#method-format}{\code{CallbackOutput$format()}}
\item \href{#method-clone}{\code{CallbackOutput$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize CallbackOutput class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CallbackOutput$new(output_name, output_type = CallbackOutputTypeEnum$String)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{output_name}}{(str): The output name}

\item{\code{output_type}}{(CallbackOutputTypeEnum): The output type}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_request"></a>}}
\if{latex}{\out{\hypertarget{method-to_request}{}}}
\subsection{Method \code{to_request()}}{
Get the request structure for workflow service calls.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CallbackOutput$to_request()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-expr"></a>}}
\if{latex}{\out{\hypertarget{method-expr}{}}}
\subsection{Method \code{expr()}}{
The 'Get' expression dict for a `CallbackOutput`.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CallbackOutput$expr(step_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{step_name}}{(str): The name of the step the callback step associated
with this output belongs to.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-format"></a>}}
\if{latex}{\out{\hypertarget{method-format}{}}}
\subsection{Method \code{format()}}{
format class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CallbackOutput$format()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CallbackOutput$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
