% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_callback_step.R
\docType{data}
\name{CallbackOutputTypeEnum}
\alias{CallbackOutputTypeEnum}
\title{Workflow CallbackOuputTypeEnum class}
\format{
An object of class \code{Enum} (inherits from \code{environment}) of length 4.
}
\usage{
CallbackOutputTypeEnum
}
\description{
CallbackOutput type enum.
}
\keyword{datasets}
