% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_callback_step.R
\name{CallbackStep}
\alias{CallbackStep}
\title{Callback step for workflow.}
\description{
Callback step for workflow.

Callback step for workflow.
}
\section{Super classes}{
\code{\link[sagemaker.workflow:Entity]{sagemaker.workflow::Entity}} -> \code{\link[sagemaker.workflow:Step]{sagemaker.workflow::Step}} -> \code{CallbackStep}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{sqs_queue_url}}{An SQS queue URL for receiving callback messages.}

\item{\code{inputs}}{Input arguments that will be provided in the SQS message body of callback messages}

\item{\code{outputs}}{Outputs that can be provided when completing a callback.}

\item{\code{cache_config}}{A list of step names this `TransformStep`}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{arguments}}{The arguments dict that is used to define the callback step}

\item{\code{properties}}{A Properties object representing the output parameters of the callback step.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CallbackStep$new()}}
\item \href{#method-to_request}{\code{CallbackStep$to_request()}}
\item \href{#method-clone}{\code{CallbackStep$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="sagemaker.workflow" data-topic="Step" data-id="add_depends_on">}\href{../../sagemaker.workflow/html/Step.html#method-add_depends_on}{\code{sagemaker.workflow::Step$add_depends_on()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sagemaker.workflow" data-topic="Step" data-id="format">}\href{../../sagemaker.workflow/html/Step.html#method-format}{\code{sagemaker.workflow::Step$format()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Constructs a CallbackStep.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CallbackStep$new(
  name,
  sqs_queue_url,
  inputs,
  outputs,
  cache_config = NULL,
  depends_on = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{(str): The name of the callback step.}

\item{\code{sqs_queue_url}}{(str): An SQS queue URL for receiving callback messages.}

\item{\code{inputs}}{(dict): Input arguments that will be provided
in the SQS message body of callback messages.}

\item{\code{outputs}}{(List[CallbackOutput]): Outputs that can be provided when completing a callback.}

\item{\code{cache_config}}{(CacheConfig):  A `CacheConfig` instance.}

\item{\code{depends_on}}{(List[str]): A list of step names this `TransformStep`
depends on}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_request"></a>}}
\if{latex}{\out{\hypertarget{method-to_request}{}}}
\subsection{Method \code{to_request()}}{
Updates the dictionary with cache configuration.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CallbackStep$to_request()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CallbackStep$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
