% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_check_job_config.R
\name{CheckJobConfig}
\alias{CheckJobConfig}
\title{CheckJobConfig class}
\description{
Check job config for QualityCheckStep and ClarifyCheckStep
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CheckJobConfig$new()}}
\item \href{#method-.generate_model_monitor}{\code{CheckJobConfig$.generate_model_monitor()}}
\item \href{#method-format}{\code{CheckJobConfig$format()}}
\item \href{#method-clone}{\code{CheckJobConfig$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Constructs a CheckJobConfig instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CheckJobConfig$new(
  role,
  instance_count = 1,
  instance_type = "ml.m5.xlarge",
  volume_size_in_gb = 30,
  volume_kms_key = NULL,
  output_kms_key = NULL,
  max_runtime_in_seconds = NULL,
  base_job_name = NULL,
  sagemaker_session = NULL,
  env = NULL,
  tags = NULL,
  network_config = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{role}}{(str): An AWS IAM role. The Amazon SageMaker jobs use this role.}

\item{\code{instance_count}}{(int): The number of instances to run the jobs with (default: 1).}

\item{\code{instance_type}}{(str): Type of EC2 instance to use for the job
(default: 'ml.m5.xlarge').}

\item{\code{volume_size_in_gb}}{(int): Size in GB of the EBS volume
to use for storing data during processing (default: 30).}

\item{\code{volume_kms_key}}{(str): A KMS key for the processing volume (default: None).}

\item{\code{output_kms_key}}{(str): The KMS key id for the job's outputs (default: None).}

\item{\code{max_runtime_in_seconds}}{(int): Timeout in seconds. After this amount of
time, Amazon SageMaker terminates the job regardless of its current status.
Default: 3600 if not specified}

\item{\code{base_job_name}}{(str): Prefix for the job name. If not specified,
a default name is generated based on the training image name and
current timestamp (default: None).}

\item{\code{sagemaker_session}}{(sagemaker.session.Session): Session object which
manages interactions with Amazon SageMaker APIs and any other
AWS services needed (default: None). If not specified, one is
created using the default AWS configuration chain.}

\item{\code{env}}{(dict): Environment variables to be passed to the job (default: None).}

\item{\code{tags}}{([dict]): List of tags to be passed to the job (default: None).}

\item{\code{network_config}}{(sagemaker.network.NetworkConfig): A NetworkConfig
object that configures network isolation, encryption of
inter-container traffic, security group IDs, and subnets (default: None).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-.generate_model_monitor"></a>}}
\if{latex}{\out{\hypertarget{method-.generate_model_monitor}{}}}
\subsection{Method \code{.generate_model_monitor()}}{
Generates a ModelMonitor object
             Generates a ModelMonitor object with required config attributes for
             QualityCheckStep and ClarifyCheckStep
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CheckJobConfig$.generate_model_monitor(mm_type)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mm_type}}{(str): The subclass type of ModelMonitor object.
A valid mm_type should be one of the following: "DefaultModelMonitor",
"ModelQualityMonitor", "ModelBiasMonitor", "ModelExplainabilityMonitor"}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
sagemaker.model_monitor.ModelMonitor or None if the mm_type is not valid
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-format"></a>}}
\if{latex}{\out{\hypertarget{method-format}{}}}
\subsection{Method \code{format()}}{
Format class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CheckJobConfig$format()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CheckJobConfig$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
