% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_clarify_check_step.R
\name{ClarifyCheckConfig}
\alias{ClarifyCheckConfig}
\title{Clarify Check Config}
\description{
Clarify Check Config

Clarify Check Config
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data_config}}{Config of the input/output data.}

\item{\code{kms_key}}{The ARN of the KMS key that is used to encrypt the
user code file}

\item{\code{monitoring_analysis_config_uri}}{The uri of monitoring analysis config.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ClarifyCheckConfig$new()}}
\item \href{#method-format}{\code{ClarifyCheckConfig$format()}}
\item \href{#method-clone}{\code{ClarifyCheckConfig$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize ClarifyCheckConfig class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClarifyCheckConfig$new(
  data_config,
  kms_key = NULL,
  monitoring_analysis_config_uri = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_config}}{(DataConfig): Config of the input/output data.}

\item{\code{kms_key}}{(str): The ARN of the KMS key that is used to encrypt the
user code file (default: None).
This field CANNOT be any of PipelineNonPrimitiveInputTypes.}

\item{\code{monitoring_analysis_config_uri}}{(str): The uri of monitoring analysis config.
This field does not take input.
It will be generated once uploading the created analysis config file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-format"></a>}}
\if{latex}{\out{\hypertarget{method-format}{}}}
\subsection{Method \code{format()}}{
format class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClarifyCheckConfig$format()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClarifyCheckConfig$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
