% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_conditions.R
\name{Condition}
\alias{Condition}
\title{Workflow Condition class}
\description{
Abstract Condition entity.
}
\section{Super class}{
\code{\link[sagemaker.workflow:Entity]{sagemaker.workflow::Entity}} -> \code{Condition}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{condition_type}}{The type of condition.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Condition$new()}}
\item \href{#method-clone}{\code{Condition$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="sagemaker.workflow" data-topic="Entity" data-id="format">}\href{../../sagemaker.workflow/html/Entity.html#method-format}{\code{sagemaker.workflow::Entity$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sagemaker.workflow" data-topic="Entity" data-id="to_request">}\href{../../sagemaker.workflow/html/Entity.html#method-to_request}{\code{sagemaker.workflow::Entity$to_request()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize Condition class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Condition$new(condition_type = enum_items(ConditionTypeEnum))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{condition_type}}{(ConditionTypeEnum): The type of condition.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Condition$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
