% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_conditions.R
\name{ConditionComparison}
\alias{ConditionComparison}
\title{Workflow ConditionComparison class}
\description{
Generic comparison condition that can be used to derive specific
             condition comparisons.
}
\section{Super classes}{
\code{\link[R6sagemaker.workflow:Entity]{R6sagemaker.workflow::Entity}} -> \code{\link[R6sagemaker.workflow:Condition]{R6sagemaker.workflow::Condition}} -> \code{ConditionComparison}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{left}}{The execution variable, parameter, or property to use in the comparison.}

\item{\code{right}}{The execution variable, parameter, property, or Python primitive value to compare to.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ConditionComparison$new()}}
\item \href{#method-to_request}{\code{ConditionComparison$to_request()}}
\item \href{#method-clone}{\code{ConditionComparison$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.workflow" data-topic="Entity" data-id="format">}\href{../../R6sagemaker.workflow/html/Entity.html#method-format}{\code{R6sagemaker.workflow::Entity$format()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize ConditionComparison Class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConditionComparison$new(
  condition_type = enum_items(ConditionTypeEnum),
  left,
  right
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{condition_type}}{(ConditionTypeEnum): The type of condition.}

\item{\code{left}}{(ConditionValueType): The execution variable, parameter, or
property to use in the comparison.}

\item{\code{right}}{(Union[ConditionValueType, PrimitiveType]): The execution variable,
parameter, property, or Python primitive value to compare to.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_request"></a>}}
\if{latex}{\out{\hypertarget{method-to_request}{}}}
\subsection{Method \code{to_request()}}{
Get the request structure for workflow service calls.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConditionComparison$to_request()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConditionComparison$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
