% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_conditions.R
\name{ConditionGreaterThanOrEqualTo}
\alias{ConditionGreaterThanOrEqualTo}
\title{Workflow ConditionGreaterThanOrEqualTo class}
\description{
A condition for greater than or equal to comparisons.
}
\section{Super classes}{
\code{\link[sagemaker.workflow:Entity]{sagemaker.workflow::Entity}} -> \code{\link[sagemaker.workflow:Condition]{sagemaker.workflow::Condition}} -> \code{\link[sagemaker.workflow:ConditionComparison]{sagemaker.workflow::ConditionComparison}} -> \code{ConditionGreaterThanOrEqualTo}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ConditionGreaterThanOrEqualTo$new()}}
\item \href{#method-clone}{\code{ConditionGreaterThanOrEqualTo$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="sagemaker.workflow" data-topic="Entity" data-id="format">}\href{../../sagemaker.workflow/html/Entity.html#method-format}{\code{sagemaker.workflow::Entity$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sagemaker.workflow" data-topic="ConditionComparison" data-id="to_request">}\href{../../sagemaker.workflow/html/ConditionComparison.html#method-to_request}{\code{sagemaker.workflow::ConditionComparison$to_request()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Construct of ConditionGreaterThanOrEqualTo for greater than or equal to comparisons.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConditionGreaterThanOrEqualTo$new(left, right)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{left}}{(ConditionValueType): The execution variable, parameter,
or property to use in the comparison.}

\item{\code{right}}{(Union[ConditionValueType, PrimitiveType]): The execution
variable, parameter, property, or Python primitive value to compare to.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConditionGreaterThanOrEqualTo$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
