% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_condition_step.R
\name{ConditionStep}
\alias{ConditionStep}
\title{Workflow ConditionStep class}
\description{
Conditional step for pipelines to support conditional branching in the execution of steps.
}
\section{Super classes}{
\code{\link[R6sagemaker.workflow:Entity]{R6sagemaker.workflow::Entity}} -> \code{\link[R6sagemaker.workflow:Step]{R6sagemaker.workflow::Step}} -> \code{ConditionStep}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{conditions}}{The name of the step.}

\item{\code{if_steps}}{A list of `sagemaker.workflow.steps.Step` and
`sagemaker.workflow.step_collections.StepCollection` instances}

\item{\code{else_steps}}{A list of `sagemaker.workflow.steps.Step` and
`sagemaker.workflow.step_collections.StepCollection` instances}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{arguments}}{The arguments dict that is used to define the conditional branching in the pipeline.}

\item{\code{properties}}{A simple Properties object with `Outcome` as the only property}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ConditionStep$new()}}
\item \href{#method-clone}{\code{ConditionStep$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.workflow" data-topic="Step" data-id="add_depends_on">}\href{../../R6sagemaker.workflow/html/Step.html#method-add_depends_on}{\code{R6sagemaker.workflow::Step$add_depends_on()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.workflow" data-topic="Step" data-id="format">}\href{../../R6sagemaker.workflow/html/Step.html#method-format}{\code{R6sagemaker.workflow::Step$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.workflow" data-topic="Step" data-id="to_request">}\href{../../R6sagemaker.workflow/html/Step.html#method-to_request}{\code{R6sagemaker.workflow::Step$to_request()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Construct a ConditionStep for pipelines to support conditional branching.
             If all of the conditions in the condition list evaluate to True, the `if_steps` are
             marked as ready for execution. Otherwise, the `else_steps` are marked as ready for
             execution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConditionStep$new(
  name,
  depends_on = NULL,
  conditions = NULL,
  if_steps = NULL,
  else_steps = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{(str): The name of the step.}

\item{\code{depends_on}}{(List[str]): The list of step names the current step depends on}

\item{\code{conditions}}{(List[Condition]): A list of `sagemaker.workflow.conditions.Condition`
instances.}

\item{\code{if_steps}}{(List[Union[Step, StepCollection]]): A list of `sagemaker.workflow.steps.Step`
and `sagemaker.workflow.step_collections.StepCollection` instances that are
marked as ready for execution if the list of conditions evaluates to True.}

\item{\code{else_steps}}{(List[Union[Step, StepCollection]]): A list of `sagemaker.workflow.steps.Step`
and `sagemaker.workflow.step_collections.StepCollection` instances that are
marked as ready for execution if the list of conditions evaluates to False.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConditionStep$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
