% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_conditions.R
\docType{data}
\name{ConditionTypeEnum}
\alias{ConditionTypeEnum}
\title{Workflow CondiitionType class}
\format{
An object of class \code{Enum} (inherits from \code{environment}) of length 8.
}
\usage{
ConditionTypeEnum
}
\description{
Condition type enum.
}
\keyword{datasets}
