% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_steps.R
\name{ConfigurableRetryStep}
\alias{ConfigurableRetryStep}
\title{ConfigurableRetryStep class}
\description{
ConfigurableRetryStep step for workflow.
}
\section{Super classes}{
\code{\link[sagemaker.workflow:Entity]{sagemaker.workflow::Entity}} -> \code{\link[sagemaker.workflow:Step]{sagemaker.workflow::Step}} -> \code{ConfigurableRetryStep}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ConfigurableRetryStep$new()}}
\item \href{#method-add_retry_policy}{\code{ConfigurableRetryStep$add_retry_policy()}}
\item \href{#method-to_request}{\code{ConfigurableRetryStep$to_request()}}
\item \href{#method-clone}{\code{ConfigurableRetryStep$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="sagemaker.workflow" data-topic="Step" data-id="add_depends_on">}\href{../../sagemaker.workflow/html/Step.html#method-add_depends_on}{\code{sagemaker.workflow::Step$add_depends_on()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sagemaker.workflow" data-topic="Step" data-id="format">}\href{../../sagemaker.workflow/html/Step.html#method-format}{\code{sagemaker.workflow::Step$format()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize ConfigurableRetryStep class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfigurableRetryStep$new(
  name,
  step_type = enum_items(StepTypeEnum),
  display_name = NULL,
  description = NULL,
  depends_on = NULL,
  retry_policies = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{(str): The name of the step.}

\item{\code{step_type}}{(StepTypeEnum): The type of the step.}

\item{\code{display_name}}{(str): The display name of the step.}

\item{\code{description}}{(str): The description of the step.}

\item{\code{depends_on}}{(List[str] or List[Step]): The list of step names or step
instances the current step depends on}

\item{\code{retry_policies}}{(List[RetryPolicy]): The custom retry policy configuration}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_retry_policy"></a>}}
\if{latex}{\out{\hypertarget{method-add_retry_policy}{}}}
\subsection{Method \code{add_retry_policy()}}{
Add a retry policy to the current step retry policies list.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfigurableRetryStep$add_retry_policy(retry_policy)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{retry_policy}}{: Placeholder}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_request"></a>}}
\if{latex}{\out{\hypertarget{method-to_request}{}}}
\subsection{Method \code{to_request()}}{
Gets the request structure for ConfigurableRetryStep
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfigurableRetryStep$to_request()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfigurableRetryStep$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
