% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_clarify_check_step.R
\name{DataBiasCheckConfig}
\alias{DataBiasCheckConfig}
\title{Data Bias Check Config}
\description{
Data Bias Check Config

Data Bias Check Config
}
\section{Super class}{
\code{\link[sagemaker.workflow:ClarifyCheckConfig]{sagemaker.workflow::ClarifyCheckConfig}} -> \code{DataBiasCheckConfig}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data_bias_config}}{Config of sensitive groups}

\item{\code{methods}}{Selector of a subset of potential metrics}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DataBiasCheckConfig$new()}}
\item \href{#method-clone}{\code{DataBiasCheckConfig$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="sagemaker.workflow" data-topic="ClarifyCheckConfig" data-id="format">}\href{../../sagemaker.workflow/html/ClarifyCheckConfig.html#method-format}{\code{sagemaker.workflow::ClarifyCheckConfig$format()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize DataBiasCheckConfig class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataBiasCheckConfig$new(data_bias_config, methods = "all", ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_bias_config}}{(BiasConfig): Config of sensitive groups.}

\item{\code{methods}}{(str or list[str]): Selector of a subset of potential metrics:
\itemize{
    \item{"CI" \url{https://docs.aws.amazon.com/sagemaker/latest/dg/clarify-bias-metric-class-imbalance.html}},
    \item{"DPL" \url{https://docs.aws.amazon.com/sagemaker/latest/dg/clarify-data-bias-metric-true-label-imbalance.html}},
    \item{"KL" \url{https://docs.aws.amazon.com/sagemaker/latest/dg/clarify-data-bias-metric-kl-divergence.html}},
    \item{"JS" \url{https://docs.aws.amazon.com/sagemaker/latest/dg/clarify-data-bias-metric-jensen-shannon-divergence.html}},
    \item{"LP" \url{https://docs.aws.amazon.com/sagemaker/latest/dg/clarify-data-bias-metric-lp-norm.html}},
    \item{"TVD" \url{https://docs.aws.amazon.com/sagemaker/latest/dg/clarify-data-bias-metric-total-variation-distance.html}},
    \item{"KS" \url{https://docs.aws.amazon.com/sagemaker/latest/dg/clarify-data-bias-metric-kolmogorov-smirnov.html}},
    \item{"CDDL" \url{https://docs.aws.amazon.com/sagemaker/latest/dg/clarify-data-bias-metric-cddl.html}}
}
        Defaults to computing all.
        This field CANNOT be any of PipelineNonPrimitiveInputTypes.}

\item{\code{...}}{: Parameters from ClarifyCheckConfig}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataBiasCheckConfig$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
