% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_quality_check_step.R
\name{DataQualityCheckConfig}
\alias{DataQualityCheckConfig}
\title{DataQualityCheckConfig class}
\description{
Data Quality Check Config.
}
\section{Super class}{
\code{\link[sagemaker.workflow:QualityCheckConfig]{sagemaker.workflow::QualityCheckConfig}} -> \code{DataQualityCheckConfig}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{record_preprocessor_script}}{(str): The path to the record preprocessor script
(default: None).
This can be a local path or an S3 uri string
but CANNOT be any of PipelineNonPrimitiveInputTypes.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-clone}{\code{DataQualityCheckConfig$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataQualityCheckConfig$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
