% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangler_processing.R
\name{DataWranglerProcessor}
\alias{DataWranglerProcessor}
\title{DataWranglerProcessor class}
\description{
Handles Amazon SageMaker DataWrangler tasks
}
\section{Super class}{
\code{\link[R6sagemaker.common:Processor]{R6sagemaker.common::Processor}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DataWranglerProcessor$new()}}
\item \href{#method-clone}{\code{DataWranglerProcessor$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.common" data-topic="Processor" data-id="format">}\href{../../R6sagemaker.common/html/Processor.html#method-format}{\code{R6sagemaker.common::Processor$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.common" data-topic="Processor" data-id="run">}\href{../../R6sagemaker.common/html/Processor.html#method-run}{\code{R6sagemaker.common::Processor$run()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initializes a ``Processor`` instance.
             The ``Processor`` handles Amazon SageMaker Processing tasks.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataWranglerProcessor$new(
  role,
  data_wrangler_flow_source,
  instance_count,
  instance_type,
  volume_size_in_gb = 30L,
  volume_kms_key = NULL,
  output_kms_key = NULL,
  max_runtime_in_seconds = NULL,
  base_job_name = NULL,
  sagemaker_session = NULL,
  env = NULL,
  tags = NULL,
  network_config = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{role}}{(str): An AWS IAM role name or ARN. Amazon SageMaker Processing
uses this role to access AWS resources, such as
data stored in Amazon S3.}

\item{\code{data_wrangler_flow_source}}{(str): The source of the DaraWrangler flow which will be
used for the DataWrangler job. If a local path is provided, it will automatically
be uploaded to S3 under:
"s3://<default-bucket-name>/<job-name>/input/<input-name>".}

\item{\code{instance_count}}{(int): The number of instances to run
a processing job with.}

\item{\code{instance_type}}{(str): The type of EC2 instance to use for
processing, for example, 'ml.c4.xlarge'.}

\item{\code{volume_size_in_gb}}{(int): Size in GB of the EBS volume
to use for storing data during processing (default: 30).}

\item{\code{volume_kms_key}}{(str): A KMS key for the processing
volume (default: None).}

\item{\code{output_kms_key}}{(str): The KMS key ID for processing job outputs (default: None).}

\item{\code{max_runtime_in_seconds}}{(int): Timeout in seconds (default: None).
After this amount of time, Amazon SageMaker terminates the job,
regardless of its current status. If `max_runtime_in_seconds` is not
specified, the default value is 24 hours.}

\item{\code{base_job_name}}{(str): Prefix for processing job name. If not specified,
the processor generates a default job name, based on the
processing image name and current timestamp.}

\item{\code{sagemaker_session}}{(:class:`~sagemaker.session.Session`):
Session object which manages interactions with Amazon SageMaker and
any other AWS services needed. If not specified, the processor creates
one using the default AWS configuration chain.}

\item{\code{env}}{(dict[str, str]): Environment variables to be passed to
the processing jobs (default: None).}

\item{\code{tags}}{(list[dict]): List of tags to be passed to the processing job
(default: None). For more, see
\url{https://docs.aws.amazon.com/sagemaker/latest/dg/API_Tag.html}.}

\item{\code{network_config}}{(:class:`~sagemaker.network.NetworkConfig`):
A :class:`~sagemaker.network.NetworkConfig`
object that configures network isolation, encryption of
inter-container traffic, security group IDs, and subnets.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataWranglerProcessor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
