% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_airflow.R
\name{deploy_config_from_estimator}
\alias{deploy_config_from_estimator}
\title{Export Airflow deploy config from a SageMaker estimator}
\usage{
deploy_config_from_estimator(
  estimator,
  task_id,
  task_type,
  initial_instance_count,
  instance_type,
  model_name = NULL,
  endpoint_name = NULL,
  tags = NULL,
  ...
)
}
\arguments{
\item{estimator}{(sagemaker.model.EstimatorBase): The SageMaker estimator to
export Airflow config from. It has to be an estimator associated
with a training job.}

\item{task_id}{(str): The task id of any
airflow.contrib.operators.SageMakerTrainingOperator or
airflow.contrib.operators.SageMakerTuningOperator that generates
training jobs in the DAG. The endpoint config is built based on the
training job generated in this operator.}

\item{task_type}{(str): Whether the task is from SageMakerTrainingOperator or
SageMakerTuningOperator. Values can be 'training', 'tuning' or None
(which means training job is not from any task).}

\item{initial_instance_count}{(int): Minimum number of EC2 instances to deploy
to an endpoint for prediction.}

\item{instance_type}{(str): Type of EC2 instance to deploy to an endpoint for
prediction, for example, 'ml.c4.xlarge'.}

\item{model_name}{(str): Name to use for creating an Amazon SageMaker model. If
not specified, one will be generated.}

\item{endpoint_name}{(str): Name to use for creating an Amazon SageMaker
endpoint. If not specified, the name of the SageMaker model is used.}

\item{tags}{(list[dict]): List of tags for labeling a training job. For more,
see https://docs.aws.amazon.com/sagemaker/latest/dg/API_Tag.html.}

\item{...}{: Passed to invocation of ``create_model()``. Implementations
may customize ``create_model()`` to accept ``**kwargs`` to customize
model creation during deploy. For more, see the implementation docs.}
}
\value{
dict: Deploy config that can be directly used by
             SageMakerEndpointOperator in Airflow.
}
\description{
Export Airflow deploy config from a SageMaker estimator
}
