% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_pipeline.R
\name{.PipeLineExecution}
\alias{.PipeLineExecution}
\title{}
\description{
Internal class for encapsulating pipeline execution instances.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{arn}}{The arn of the pipeline execution}

\item{\code{sagemaker_session}}{Session object which manages interactions with Amazon SageMaker}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{.PipeLineExecution$new()}}
\item \href{#method-stop}{\code{.PipeLineExecution$stop()}}
\item \href{#method-describe}{\code{.PipeLineExecution$describe()}}
\item \href{#method-list_steps}{\code{.PipeLineExecution$list_steps()}}
\item \href{#method-wait}{\code{.PipeLineExecution$wait()}}
\item \href{#method-format}{\code{.PipeLineExecution$format()}}
\item \href{#method-clone}{\code{.PipeLineExecution$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize .PipeLineExecution class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.PipeLineExecution$new(arn, sagemaker_session = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{arn}}{(str): The arn of the pipeline execution.}

\item{\code{sagemaker_session}}{(sagemaker.session.Session): Session object which
manages interactions with Amazon SageMaker APIs and any other
AWS services needed. If not specified, the
pipeline creates one using the default AWS configuration chain.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-stop"></a>}}
\if{latex}{\out{\hypertarget{method-stop}{}}}
\subsection{Method \code{stop()}}{
Stops a pipeline execution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.PipeLineExecution$stop()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-describe"></a>}}
\if{latex}{\out{\hypertarget{method-describe}{}}}
\subsection{Method \code{describe()}}{
Describes a pipeline execution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.PipeLineExecution$describe()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Information about the pipeline execution. See
             `boto3 client describe_pipeline_execution
             \url{https://boto3.amazonaws.com/v1/documentation/api/latest/reference/services/sagemaker.html#SageMaker.Client.describe_pipeline_execution}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_steps"></a>}}
\if{latex}{\out{\hypertarget{method-list_steps}{}}}
\subsection{Method \code{list_steps()}}{
Describes a pipeline execution's steps.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.PipeLineExecution$list_steps()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Information about the steps of the pipeline execution. See
             `boto3 client list_pipeline_execution_steps
             \url{https://boto3.amazonaws.com/v1/documentation/api/latest/reference/services/sagemaker.html#SageMaker.Client.list_pipeline_execution_steps}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-wait"></a>}}
\if{latex}{\out{\hypertarget{method-wait}{}}}
\subsection{Method \code{wait()}}{
Waits for a pipeline execution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.PipeLineExecution$wait(delay = 30, max_attempts = 60)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{delay}}{(int): The polling interval. (Defaults to 30 seconds)}

\item{\code{max_attempts}}{(int): The maximum number of polling attempts.
(Defaults to 60 polling attempts)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-format"></a>}}
\if{latex}{\out{\hypertarget{method-format}{}}}
\subsection{Method \code{format()}}{
format class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.PipeLineExecution$format()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.PipeLineExecution$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
