% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow__utils.R
\name{.RepackModelStep}
\alias{.RepackModelStep}
\title{Workflow .RepackModelStep class}
\description{
Repacks model artifacts with inference entry point.
}
\keyword{internal}
\section{Super classes}{
\code{\link[R6sagemaker.workflow:Entity]{R6sagemaker.workflow::Entity}} -> \code{\link[R6sagemaker.workflow:Step]{R6sagemaker.workflow::Step}} -> \code{\link[R6sagemaker.workflow:TrainingStep]{R6sagemaker.workflow::TrainingStep}} -> \code{.RepackModelStep}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{The name of the training step.}

\item{\code{step_type}}{The type of the step with value `StepTypeEnum.Training`.}

\item{\code{estimator}}{A `sagemaker.estimator.EstimatorBase` instance.}

\item{\code{inputs}}{A `sagemaker.inputs.TrainingInput` instance. Defaults to `None`.}

\item{\code{sagemaker_session}}{`R6sagemaker.common::Session` class}

\item{\code{role}}{execution role}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{.RepackModelStep$new()}}
\item \href{#method-clone}{\code{.RepackModelStep$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.workflow" data-topic="Step" data-id="add_depends_on">}\href{../../R6sagemaker.workflow/html/Step.html#method-add_depends_on}{\code{R6sagemaker.workflow::Step$add_depends_on()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.workflow" data-topic="Step" data-id="format">}\href{../../R6sagemaker.workflow/html/Step.html#method-format}{\code{R6sagemaker.workflow::Step$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.workflow" data-topic="TrainingStep" data-id="to_request">}\href{../../R6sagemaker.workflow/html/TrainingStep.html#method-to_request}{\code{R6sagemaker.workflow::TrainingStep$to_request()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Constructs a TrainingStep, given an `EstimatorBase` instance.
             In addition to the estimator instance, the other arguments are those that are supplied to
             the `fit` method of the `sagemaker.estimator.Estimator`.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.RepackModelStep$new(
  name,
  sagemaker_session,
  role,
  model_data,
  entry_point,
  source_dir = NULL,
  dependencies = NULL,
  depends_on = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{(str): The name of the training step.}

\item{\code{sagemaker_session}}{(EstimatorBase): A `sagemaker.estimator.EstimatorBase` instance.}

\item{\code{role}}{(str) : placeholder}

\item{\code{model_data}}{: placeholder}

\item{\code{entry_point}}{: placeholder}

\item{\code{source_dir}}{: placeholder}

\item{\code{dependencies}}{: placeholder}

\item{\code{depends_on}}{: placeholder}

\item{\code{...}}{: additional parameters passed `R6sagemaker.mlframework::SKLearn` class}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.RepackModelStep$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
