% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_emr_step.R
\name{EMRStep}
\alias{EMRStep}
\title{EMRStep class}
\description{
EMR step for workflow.
}
\section{Super classes}{
\code{\link[sagemaker.workflow:Entity]{sagemaker.workflow::Entity}} -> \code{\link[sagemaker.workflow:Step]{sagemaker.workflow::Step}} -> \code{EMRStep}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{arguments}}{The arguments dict that is used to call `AddJobFlowSteps`.
NOTE: The AddFlowJobSteps request is not quite the args list that workflow needs.
The Name attribute in AddJobFlowSteps cannot be passed; it will be set during runtime.
In addition to that, we will also need to include emr job inputs and output config.}

\item{\code{properties}}{A Properties object representing the EMR DescribeStepResponse model}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{EMRStep$new()}}
\item \href{#method-to_request}{\code{EMRStep$to_request()}}
\item \href{#method-clone}{\code{EMRStep$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="sagemaker.workflow" data-topic="Step" data-id="add_depends_on">}\href{../../sagemaker.workflow/html/Step.html#method-add_depends_on}{\code{sagemaker.workflow::Step$add_depends_on()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sagemaker.workflow" data-topic="Step" data-id="format">}\href{../../sagemaker.workflow/html/Step.html#method-format}{\code{sagemaker.workflow::Step$format()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Constructs a EMRStep.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EMRStep$new(
  name,
  display_name,
  description,
  cluster_id,
  step_config,
  depends_on = NULL,
  cache_config = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{(str): The name of the EMR step.}

\item{\code{display_name}}{(str): The display name of the EMR step.}

\item{\code{description}}{(str): The description of the EMR step.}

\item{\code{cluster_id}}{(str): The ID of the running EMR cluster.}

\item{\code{step_config}}{(EMRStepConfig): One StepConfig to be executed by the job flow.}

\item{\code{depends_on}}{(List[str]): A list of step names this `sagemaker.workflow.steps.EMRStep`
depends on}

\item{\code{cache_config}}{(CacheConfig): A `sagemaker.workflow.steps.CacheConfig` instance.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_request"></a>}}
\if{latex}{\out{\hypertarget{method-to_request}{}}}
\subsection{Method \code{to_request()}}{
Updates the dictionary with cache configuration.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EMRStep$to_request()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EMRStep$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
