% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_emr_step.R
\name{EMRStepConfig}
\alias{EMRStepConfig}
\title{EMRStepConfig class}
\description{
Config for a Hadoop Jar step
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{jar}}{A path to a JAR file run during the step.}

\item{\code{args}}{A list of command line arguments}

\item{\code{main_class}}{The name of the main class in the specified Java file.}

\item{\code{properties}}{A list of key-value pairs that are set when the step runs.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{EMRStepConfig$new()}}
\item \href{#method-to_request}{\code{EMRStepConfig$to_request()}}
\item \href{#method-format}{\code{EMRStepConfig$format()}}
\item \href{#method-clone}{\code{EMRStepConfig$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a definition for input data used by an EMR cluster(job flow) step.
             See AWS documentation on the ``StepConfig`` API for more details on the parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EMRStepConfig$new(jar, args = NULL, main_class = NULL, properties = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{jar}}{(str): A path to a JAR file run during the step.}

\item{\code{args}}{(List[str]): A list of command line arguments passed to
the JAR file's main function when executed.}

\item{\code{main_class}}{(str): The name of the main class in the specified Java file.}

\item{\code{properties}}{(List(dict)): A list of key-value pairs that are set when the step runs.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_request"></a>}}
\if{latex}{\out{\hypertarget{method-to_request}{}}}
\subsection{Method \code{to_request()}}{
Convert EMRStepConfig object to request list.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EMRStepConfig$to_request()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-format"></a>}}
\if{latex}{\out{\hypertarget{method-format}{}}}
\subsection{Method \code{format()}}{
format class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EMRStepConfig$format()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EMRStepConfig$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
