% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_step_collections.R
\name{EstimatorTransformer}
\alias{EstimatorTransformer}
\title{Workflow EstimatorTransformer class}
\description{
Creates a Transformer step collection for workflow.
}
\section{Super class}{
\code{\link[R6sagemaker.workflow:StepCollection]{R6sagemaker.workflow::StepCollection}} -> \code{EstimatorTransformer}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{EstimatorTransformer$new()}}
\item \href{#method-clone}{\code{EstimatorTransformer$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.workflow" data-topic="StepCollection" data-id="format">}\href{../../R6sagemaker.workflow/html/StepCollection.html#method-format}{\code{R6sagemaker.workflow::StepCollection$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.workflow" data-topic="StepCollection" data-id="request_list">}\href{../../R6sagemaker.workflow/html/StepCollection.html#method-request_list}{\code{R6sagemaker.workflow::StepCollection$request_list()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Construct steps required for a Transformer step collection:
             An estimator-centric step collection. It models what happens in workflows
             when invoking the `transform()` method on an estimator instance:
             First, if custom
             model artifacts are required, a `_RepackModelStep` is included.
             Second, a `CreateModelStep` with the model data passed in from
             a training step or other training job output. Finally, a `TransformerStep`.
             If repacking the model artifacts is not necessary, only the
             CreateModelStep and TransformerStep are in the step collection.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EstimatorTransformer$new(
  name,
  estimator,
  model_data,
  model_inputs,
  instance_count,
  instance_type,
  transform_inputs,
  image_uri = NULL,
  predictor_cls = NULL,
  env = NULL,
  strategy = NULL,
  assemble_with = NULL,
  output_path = NULL,
  output_kms_key = NULL,
  accept = NULL,
  max_concurrent_transforms = NULL,
  max_payload = NULL,
  tags = NULL,
  volume_kms_key = NULL,
  depends_on = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{(str): The name of the Transform Step.}

\item{\code{estimator}}{: The estimator instance.}

\item{\code{model_data}}{: The S3 uri to the model data from training.}

\item{\code{model_inputs}}{: Placeholder}

\item{\code{instance_count}}{(int): The number of EC2 instances to use.}

\item{\code{instance_type}}{(str): The type of EC2 instance to use.}

\item{\code{transform_inputs}}{: Placeholder}

\item{\code{image_uri}}{: Placeholder}

\item{\code{predictor_cls}}{: Placeholder}

\item{\code{env}}{(dict): The Environment variables to be set for use during the
transform job (default: None).}

\item{\code{strategy}}{(str): The strategy used to decide how to batch records in
a single request (default: None). Valid values: 'MultiRecord'
and 'SingleRecord'.}

\item{\code{assemble_with}}{(str): How the output is assembled (default: None).
Valid values: 'Line' or 'None'.}

\item{\code{output_path}}{(str): The S3 location for saving the transform result. If
not specified, results are stored to a default bucket.}

\item{\code{output_kms_key}}{(str): Optional. A KMS key ID for encrypting the
transform output (default: None).}

\item{\code{accept}}{(str): The accept header passed by the client to
the inference endpoint. If it is supported by the endpoint,
it will be the format of the batch transform output.}

\item{\code{max_concurrent_transforms}}{: Placeholder}

\item{\code{max_payload}}{: Placeholder}

\item{\code{tags}}{: Placeholder}

\item{\code{volume_kms_key}}{: Placeholder}

\item{\code{depends_on}}{(List[str]): The list of step names the first step in
the collection depends on}

\item{\code{...}}{: pass onto model class.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EstimatorTransformer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
