% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_execution_variables.R
\name{ExecutionVariable}
\alias{ExecutionVariable}
\title{Workflow ExecutionVariable class}
\description{
Pipeline execution variables for workflow.
}
\section{Super class}{
\code{\link[R6sagemaker.workflow:Expression]{R6sagemaker.workflow::Expression}} -> \code{ExecutionVariable}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{The name of the execution variable.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{expr}}{The 'Get' expression dict for an `ExecutionVariable`.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ExecutionVariable$new()}}
\item \href{#method-clone}{\code{ExecutionVariable$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.workflow" data-topic="Expression" data-id="format">}\href{../../R6sagemaker.workflow/html/Expression.html#method-format}{\code{R6sagemaker.workflow::Expression$format()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a pipeline execution variable.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExecutionVariable$new(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{(str): The name of the execution variable.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExecutionVariable$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
