% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_execution_variables.R
\name{ExecutionVariables}
\alias{ExecutionVariables}
\title{Enum-like class for all ExecutionVariable instances.}
\description{
Considerations to move these as module-level constants should be made.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{START_DATETIME}}{}

\item{\code{CURRENT_DATETIME}}{}

\item{\code{PIPELINE_NAME}}{}

\item{\code{PIPELINE_ARN}}{}

\item{\code{PIPELINE_EXECUTION_ARN}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-format}{\code{ExecutionVariables$format()}}
\item \href{#method-clone}{\code{ExecutionVariables$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-format"></a>}}
\if{latex}{\out{\hypertarget{method-format}{}}}
\subsection{Method \code{format()}}{
format class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExecutionVariables$format()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExecutionVariables$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
