% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_pipeline.R
\name{format_start_parameters}
\alias{format_start_parameters}
\title{Formats start parameter overrides as a list of dicts.}
\usage{
format_start_parameters(parameters)
}
\arguments{
\item{parameters}{(Dict[str, Any]): A dict of named values where the keys are
the names of the parameters to pass values into.}
}
\description{
This list of dicts adheres to the request schema of:
             `{"Name": "MyParameterName", "Value": "MyValue"}`
}
