% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangler_ingestion.R
\name{generate_data_ingestion_flow_from_redshift_dataset_definition}
\alias{generate_data_ingestion_flow_from_redshift_dataset_definition}
\title{generate data ingestion flow from redshift dataset definition}
\usage{
generate_data_ingestion_flow_from_redshift_dataset_definition(
  input_name,
  redshift_dataset_definition,
  operator_version = "0.1",
  schema = NULL
)
}
\arguments{
\item{input_name}{(str): the name of the input to flow source node}

\item{redshift_dataset_definition}{(RedshiftDatasetDefinition): redshift input to flow source node}

\item{operator_version}{(str): the version of the operator}

\item{schema}{(list): the schema for the data to be ingested}
}
\value{
list: A flow only conduct data ingestion with 1-1 mapping
             output_name (str): The output name used to configure
             `sagemaker.processing.FeatureStoreOutput`
}
\description{
Generate the data ingestion only flow from redshift input
}
