% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_utilities.R
\name{hash_file}
\alias{hash_file}
\title{Get the MD5 hash of a file.}
\usage{
hash_file(path)
}
\arguments{
\item{path}{(str): The local path for the file.}
}
\value{
str: The MD5 hash of the file.
}
\description{
Get the MD5 hash of a file.
}
