% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_airflow.R
\name{input_output_list_converter}
\alias{input_output_list_converter}
\title{Converts a list of ProcessingInput or ProcessingOutput objects to a list of dicts}
\usage{
input_output_list_converter(object_list)
}
\arguments{
\item{object_list}{(list[ProcessingInput or ProcessingOutput]}
}
\value{
List of dicts
}
\description{
Converts a list of ProcessingInput or ProcessingOutput objects to a list of dicts
}
