% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_pipeline.R
\name{interpolate}
\alias{interpolate}
\title{Replaces Parameter values in a list of nested Dict[str, Any] with their workflow expression.}
\usage{
interpolate(
  request_obj,
  callback_output_to_step_map,
  lambda_output_to_step_map
)
}
\arguments{
\item{request_obj}{(RequestType): The request dict.}

\item{callback_output_to_step_map}{(list[str, str]): A dict of output name -> step name.}

\item{lambda_output_to_step_map}{(list[str, str]): Placeholder}
}
\value{
RequestType: The request dict with Parameter values replaced by their expression.
}
\description{
Replaces Parameter values in a list of nested Dict[str, Any] with their workflow expression.
}
