% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_functions.R
\name{Join}
\alias{Join}
\title{Join Class}
\description{
Join together properties.
}
\section{Super class}{
\code{\link[sagemaker.workflow:Expression]{sagemaker.workflow::Expression}} -> \code{Join}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{on}}{The primitive types and parameters to join.}

\item{\code{values}}{The string to join the values on (Defaults to "").}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{expr}}{The expression dict for a `Join` function.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Join$new()}}
\item \href{#method-clone}{\code{Join$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="sagemaker.workflow" data-topic="Expression" data-id="format">}\href{../../sagemaker.workflow/html/Expression.html#method-format}{\code{sagemaker.workflow::Expression$format()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize Join Class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Join$new(on = "", values = "")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{on}}{(str): The string to join the values on (Defaults to "").}

\item{\code{values}}{(List[Union[PrimitiveType, Parameter]]): The primitive types
and parameters to join.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Join$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
