% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_lambda_step.R
\name{LambdaOutput}
\alias{LambdaOutput}
\title{Workflow LambdaOutput class}
\description{
Output for a lambdaback step.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{output_name}}{(str): The output name}

\item{\code{output_type}}{(LambdaOutputTypeEnum): The output type}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LambdaOutput$new()}}
\item \href{#method-to_request}{\code{LambdaOutput$to_request()}}
\item \href{#method-expr}{\code{LambdaOutput$expr()}}
\item \href{#method-format}{\code{LambdaOutput$format()}}
\item \href{#method-clone}{\code{LambdaOutput$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize LambdaOutput class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LambdaOutput$new(output_name, output_type = enum_items(LambdaOutputTypeEnum))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{output_name}}{(str): The output name}

\item{\code{output_type}}{(LambdaOutputTypeEnum): The output type}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_request"></a>}}
\if{latex}{\out{\hypertarget{method-to_request}{}}}
\subsection{Method \code{to_request()}}{
Get the request structure for workflow service calls.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LambdaOutput$to_request()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-expr"></a>}}
\if{latex}{\out{\hypertarget{method-expr}{}}}
\subsection{Method \code{expr()}}{
The 'Get' expression dict for a `LambdaOutput`.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LambdaOutput$expr(step_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{step_name}}{(str): The name of the step the lambda step associated}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-format"></a>}}
\if{latex}{\out{\hypertarget{method-format}{}}}
\subsection{Method \code{format()}}{
format class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LambdaOutput$format()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LambdaOutput$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
