% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_lambda_step.R
\docType{data}
\name{LambdaOutputTypeEnum}
\alias{LambdaOutputTypeEnum}
\title{LambdaOutput type enum.}
\format{
An object of class \code{Enum} (inherits from \code{environment}) of length 4.
}
\usage{
LambdaOutputTypeEnum
}
\description{
LambdaOutput type enum.
}
\keyword{datasets}
