% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_utilities.R
\name{list_to_request}
\alias{list_to_request}
\title{Get the request structure for list of entities.}
\usage{
list_to_request(entities)
}
\arguments{
\item{entities}{(Sequence[Entity]): A list of entities.}
}
\value{
list: A request structure for a workflow service call.
}
\description{
Get the request structure for list of entities.
}
