% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_airflow.R
\name{model_config}
\alias{model_config}
\title{Export Airflow model config from a SageMaker model}
\usage{
model_config(model, instance_type = NULL, role = NULL, image_uri = NULL)
}
\arguments{
\item{model}{(sagemaker.model.Model): The Model object from which to export the Airflow config}

\item{instance_type}{(str): The EC2 instance type to deploy this Model to. For
example, 'ml.p2.xlarge'}

\item{role}{(str): The ``ExecutionRoleArn`` IAM Role ARN for the model}

\item{image_uri}{(str): An Docker image URI to use for deploying the model}
}
\value{
dict: Model config that can be directly used by SageMakerModelOperator
             in Airflow. It can also be part of the config used by
             SageMakerEndpointOperator and SageMakerTransformOperator in Airflow.
}
\description{
Export Airflow model config from a SageMaker model
}
