% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_airflow.R
\name{model_config_from_estimator}
\alias{model_config_from_estimator}
\title{Export Airflow model config from a SageMaker estimator}
\usage{
model_config_from_estimator(
  estimator,
  task_id,
  task_type,
  instance_type = NULL,
  role = NULL,
  image_uri = NULL,
  name = NULL,
  model_server_workers = NULL,
  vpc_config_override = "VPC_CONFIG_DEFAULT"
)
}
\arguments{
\item{estimator}{(sagemaker.model.EstimatorBase): The SageMaker estimator to
export Airflow config from. It has to be an estimator associated
with a training job.}

\item{task_id}{(str): The task id of any
airflow.contrib.operators.SageMakerTrainingOperator or
airflow.contrib.operators.SageMakerTuningOperator that generates
training jobs in the DAG. The model config is built based on the
training job generated in this operator.}

\item{task_type}{(str): Whether the task is from SageMakerTrainingOperator or
SageMakerTuningOperator. Values can be 'training', 'tuning' or None
(which means training job is not from any task).}

\item{instance_type}{(str): The EC2 instance type to deploy this Model to. For
example, 'ml.p2.xlarge'}

\item{role}{(str): The ``ExecutionRoleArn`` IAM Role ARN for the model}

\item{image_uri}{(str): A Docker image URI to use for deploying the model}

\item{name}{(str): Name of the model}

\item{model_server_workers}{(int): The number of worker processes used by the
inference server. If None, server will use one worker per vCPU. Only
effective when estimator is a SageMaker framework.}

\item{vpc_config_override}{(dict[str, list[str]]): Override for VpcConfig set on
the model. Default: use subnets and security groups from this Estimator.
* 'Subnets' (list[str]): List of subnet ids.
* 'SecurityGroupIds' (list[str]): List of security group ids.}
}
\value{
dict: Model config that can be directly used by SageMakerModelOperator in Airflow. It can
             also be part of the config used by SageMakerEndpointOperator in Airflow.
}
\description{
Export Airflow model config from a SageMaker estimator
}
