% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_clarify_check_step.R
\name{ModelBiasCheckConfig}
\alias{ModelBiasCheckConfig}
\title{Model Bias Check Config}
\description{
Model Bias Check Config

Model Bias Check Config
}
\section{Super class}{
\code{\link[sagemaker.workflow:ClarifyCheckConfig]{sagemaker.workflow::ClarifyCheckConfig}} -> \code{ModelBiasCheckConfig}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data_bias_config}}{Config of sensitive groups}

\item{\code{methods}}{Selector of a subset of potential metrics}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ModelBiasCheckConfig$new()}}
\item \href{#method-clone}{\code{ModelBiasCheckConfig$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize DataBiasCheckConfig class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelBiasCheckConfig$new(
  data_bias_config,
  model_config,
  model_predicted_label_config,
  methods = "all"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_bias_config}}{(BiasConfig): Config of sensitive groups.}

\item{\code{model_config}}{(ModelConfig): Config of the model and its endpoint to be created.}

\item{\code{model_predicted_label_config}}{(ModelPredictedLabelConfig): Config of how to
extract the predicted label from the model output.}

\item{\code{methods}}{(str or list[str]): Selector of a subset of potential metrics:
\itemize{
    \item{"DPPL"\url{https://docs.aws.amazon.com/sagemaker/latest/dg/clarify-post-training-bias-metric-dppl.html}},
    \item{"DI" \url{https://docs.aws.amazon.com/sagemaker/latest/dg/clarify-post-training-bias-metric-di.html}},
    \item{"DCA" \url{https://docs.aws.amazon.com/sagemaker/latest/dg/clarify-post-training-bias-metric-dca.html}},
    \item{"DCR" \url{https://docs.aws.amazon.com/sagemaker/latest/dg/clarify-post-training-bias-metric-dcr.html}},
    \item{"RD" \url{https://docs.aws.amazon.com/sagemaker/latest/dg/clarify-post-training-bias-metric-rd.html}},
    \item{"DAR" \url{https://docs.aws.amazon.com/sagemaker/latest/dg/clarify-post-training-bias-metric-dar.html}},
    \item{"DRR" \url{https://docs.aws.amazon.com/sagemaker/latest/dg/clarify-post-training-bias-metric-drr.html}},
    \item{"AD" \url{https://docs.aws.amazon.com/sagemaker/latest/dg/clarify-post-training-bias-metric-ad.html}},
    \item{"CDDPL" \url{https://docs.aws.amazon.com/sagemaker/latest/dg/clarify-post-training-bias-metric-cddpl.html}},
    \item{"TE" \url{https://docs.aws.amazon.com/sagemaker/latest/dg/clarify-post-training-bias-metric-te.html}},
    \item{"FT" \url{https://docs.aws.amazon.com/sagemaker/latest/dg/clarify-post-training-bias-metric-ft.html}}
}
        Defaults to computing all.
        This field CANNOT be any of PipelineNonPrimitiveInputTypes.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelBiasCheckConfig$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
