% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_clarify_check_step.R
\name{ModelExplainabilityCheckConfig}
\alias{ModelExplainabilityCheckConfig}
\title{Model Explainability Check Config}
\description{
Model Explainability Check Config

Model Explainability Check Config
}
\section{Super class}{
\code{\link[sagemaker.workflow:ClarifyCheckConfig]{sagemaker.workflow::ClarifyCheckConfig}} -> \code{ModelExplainabilityCheckConfig}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{model_config}}{Config of the model and its endpoint to be created}

\item{\code{explainability_config}}{Config of the specific explainability method}

\item{\code{model_scores}}{Index or JSONPath location in the model output}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ModelExplainabilityCheckConfig$new()}}
\item \href{#method-clone}{\code{ModelExplainabilityCheckConfig$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="sagemaker.workflow" data-topic="ClarifyCheckConfig" data-id="format">}\href{../../sagemaker.workflow/html/ClarifyCheckConfig.html#method-format}{\code{sagemaker.workflow::ClarifyCheckConfig$format()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize ModelExplainabilityCheckConfig class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelExplainabilityCheckConfig$new(
  model_config,
  explainability_config,
  model_scores = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model_config}}{(ModelConfig): Config of the model and its endpoint to be created.}

\item{\code{explainability_config}}{(SHAPConfig): Config of the specific explainability method.
Currently, only SHAP is supported.}

\item{\code{model_scores}}{(str or int or ModelPredictedLabelConfig): Index or JSONPath location
in the model output for the predicted scores to be explained (default: None).
This is not required if the model output is a single score. Alternatively,
an instance of ModelPredictedLabelConfig can be provided
but this field CANNOT be any of PipelineNonPrimitiveInputTypes.}

\item{\code{...}}{: Parameters from ClarifyCheckConfig}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelExplainabilityCheckConfig$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
