% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_quality_check_step.R
\name{ModelQualityCheckConfig}
\alias{ModelQualityCheckConfig}
\title{ModelQualityCheckConfig Class}
\description{
Model Quality Check Config.
}
\section{Super class}{
\code{\link[sagemaker.workflow:QualityCheckConfig]{sagemaker.workflow::QualityCheckConfig}} -> \code{ModelQualityCheckConfig}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{problem_type}}{(str or PipelineNonPrimitiveInputTypes): The type of problem of this model
quality monitoring.
Valid values are "Regression", "BinaryClassification", "MulticlassClassification".}

\item{\code{inference_attribute}}{(str or PipelineNonPrimitiveInputTypes): Index or JSONpath to
locate predicted label(s) (default: None).}

\item{\code{probability_attribute}}{(str or PipelineNonPrimitiveInputTypes): Index or JSONpath to
locate probabilities (default: None).}

\item{\code{ground_truth_attribute}}{(str or PipelineNonPrimitiveInputTypes: Index or JSONpath to
locate actual label(s) (default: None).}

\item{\code{probability_threshold_attribute}}{(str or PipelineNonPrimitiveInputTypes): Threshold to
convert probabilities to binaries (default: None).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ModelQualityCheckConfig$new()}}
\item \href{#method-clone}{\code{ModelQualityCheckConfig$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize DataQualityCheckConfig class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelQualityCheckConfig$new(
  problem_type,
  inference_attribute = NULL,
  probability_attribute = NULL,
  ground_truth_attribute = NULL,
  probability_threshold_attribute = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{problem_type}}{(str or PipelineNonPrimitiveInputTypes): The type of problem of this model
quality monitoring. Valid values are "Regression", "BinaryClassification", "MulticlassClassification".}

\item{\code{inference_attribute}}{(str or PipelineNonPrimitiveInputTypes): Index or JSONpath to
locate predicted label(s) (default: None).}

\item{\code{probability_attribute}}{(str or PipelineNonPrimitiveInputTypes): Index or JSONpath to
locate probabilities (default: None).}

\item{\code{ground_truth_attribute}}{(str or PipelineNonPrimitiveInputTypes: Index or JSONpath to
locate actual label(s) (default: None).}

\item{\code{probability_threshold_attribute}}{(str or PipelineNonPrimitiveInputTypes): Threshold to
convert probabilities to binaries (default: None).}

\item{\code{...}}{: parameter passed to QualityCheckConfig class.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelQualityCheckConfig$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
