% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_parallelism_config.R
\name{ParallelismConfiguration}
\alias{ParallelismConfiguration}
\title{ParallelismConfiguration}
\description{
Parallelism config for SageMaker pipeline
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ParallelismConfiguration$new()}}
\item \href{#method-to_request}{\code{ParallelismConfiguration$to_request()}}
\item \href{#method-format}{\code{ParallelismConfiguration$format()}}
\item \href{#method-clone}{\code{ParallelismConfiguration$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a ParallelismConfiguration
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParallelismConfiguration$new(max_parallel_execution_steps)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{max_parallel_execution_steps, }}{int:
max number of steps which could be parallelized}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_request"></a>}}
\if{latex}{\out{\hypertarget{method-to_request}{}}}
\subsection{Method \code{to_request()}}{
The request structure.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParallelismConfiguration$to_request()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-format"></a>}}
\if{latex}{\out{\hypertarget{method-format}{}}}
\subsection{Method \code{format()}}{
format class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParallelismConfiguration$format()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParallelismConfiguration$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
