% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_parameters.R
\name{ParameterInteger}
\alias{ParameterInteger}
\title{Workflow ParameterInteger class}
\description{
Pipeline integer parameter for workflow.
}
\section{Super classes}{
\code{\link[R6sagemaker.workflow:Entity]{R6sagemaker.workflow::Entity}} -> \code{\link[R6sagemaker.workflow:Parameter]{R6sagemaker.workflow::Parameter}} -> \code{ParameterInteger}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ParameterInteger$new()}}
\item \href{#method-clone}{\code{ParameterInteger$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.workflow" data-topic="Parameter" data-id="to_request">}\href{../../R6sagemaker.workflow/html/Parameter.html#method-to_request}{\code{R6sagemaker.workflow::Parameter$to_request()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a pipeline integer parameter.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParameterInteger$new(name, default_value = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{(str): The name of the parameter.}

\item{\code{default_value}}{(int): The default Python value of the parameter.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParameterInteger$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
