% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_parameters.R
\name{ParameterString}
\alias{ParameterString}
\title{Workflow ParameterString class}
\description{
Pipeline string parameter for workflow.
}
\section{Super classes}{
\code{\link[R6sagemaker.workflow:Entity]{R6sagemaker.workflow::Entity}} -> \code{\link[R6sagemaker.workflow:Parameter]{R6sagemaker.workflow::Parameter}} -> \code{ParameterString}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ParameterString$new()}}
\item \href{#method-clone}{\code{ParameterString$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.workflow" data-topic="Parameter" data-id="to_request">}\href{../../R6sagemaker.workflow/html/Parameter.html#method-to_request}{\code{R6sagemaker.workflow::Parameter$to_request()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a pipeline string parameter.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParameterString$new(name, default_value = NULL, enum_values = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{(str): The name of the parameter.}

\item{\code{default_value}}{(str): The default Python value of the parameter. Defaults to None.}

\item{\code{enum_values}}{(list): placeholder}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParameterString$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
