% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_pipeline_experiment_config.R
\name{PipelineExperimentConfig}
\alias{PipelineExperimentConfig}
\title{Workflow PipelineExperimentConfig class}
\description{
Experiment config for SageMaker pipeline.
}
\examples{

## ------------------------------------------------
## Method `PipelineExperimentConfig$new`
## ------------------------------------------------

# Use pipeline name as the experiment name and pipeline execution id as the trial name::
PipeLineExperimentConfig$new(
     ExecutionVariables$PIPELINE_NAME, ExecutionVariables$PIPELINE_EXECUTION_ID)
# Use a customized experiment name and pipeline execution id as the trial name::
PipeLineExperimentConfig$new(
     'MyExperiment', ExecutionVariables$PIPELINE_EXECUTION_ID)
}
\section{Super class}{
\code{\link[sagemaker.workflow:Entity]{sagemaker.workflow::Entity}} -> \code{PipelineExperimentConfig}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PipelineExperimentConfig$new()}}
\item \href{#method-to_request}{\code{PipelineExperimentConfig$to_request()}}
\item \href{#method-clone}{\code{PipelineExperimentConfig$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="sagemaker.workflow" data-topic="Entity" data-id="format">}\href{../../sagemaker.workflow/html/Entity.html#method-format}{\code{sagemaker.workflow::Entity$format()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a PipelineExperimentConfig
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PipelineExperimentConfig$new(experiment_name, trial_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{experiment_name}}{(Union[str, Parameter, ExecutionVariable, Expression]):
the name of the experiment that will be created.}

\item{\code{trial_name}}{(Union[str, Parameter, ExecutionVariable, Expression]):
the name of the trial that will be created.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# Use pipeline name as the experiment name and pipeline execution id as the trial name::
PipeLineExperimentConfig$new(
     ExecutionVariables$PIPELINE_NAME, ExecutionVariables$PIPELINE_EXECUTION_ID)
# Use a customized experiment name and pipeline execution id as the trial name::
PipeLineExperimentConfig$new(
     'MyExperiment', ExecutionVariables$PIPELINE_EXECUTION_ID)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_request"></a>}}
\if{latex}{\out{\hypertarget{method-to_request}{}}}
\subsection{Method \code{to_request()}}{
Returns: the request structure.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PipelineExperimentConfig$to_request()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PipelineExperimentConfig$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
