% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_pipeline_experiment_config.R
\docType{data}
\name{PipelineExperimentConfigProperties}
\alias{PipelineExperimentConfigProperties}
\title{Workflow PipelineExperimentConfigProperties enum like class}
\format{
An object of class \code{Enum} (inherits from \code{environment}) of length 2.
}
\usage{
PipelineExperimentConfigProperties
}
\description{
Enum-like class for all pipeline experiment config property references.
}
\keyword{datasets}
