% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_airflow.R
\name{prepare_amazon_algorithm_estimator}
\alias{prepare_amazon_algorithm_estimator}
\title{Sets up amazon algorithm estimator.}
\usage{
prepare_amazon_algorithm_estimator(estimator, inputs, mini_batch_size = NULL)
}
\arguments{
\item{estimator}{(sagemaker.amazon.amazon_estimator.AmazonAlgorithmEstimatorBase): An estimator
for a built-in Amazon algorithm to get information from and update.}

\item{inputs}{: The training data.
* (sagemaker.amazon.amazon_estimator.RecordSet) - A collection of
Amazon :class:~`Record` objects serialized and stored in S3. For
use with an estimator for an Amazon algorithm.
* (list[sagemaker.amazon.amazon_estimator.RecordSet]) - A list of
:class:~`sagemaker.amazon.amazon_estimator.RecordSet` objects,
where each instance is a different channel of training data.}

\item{mini_batch_size}{(numeric):}
}
\description{
This is done by adding the required `feature_dim` hyperparameter from training data.
}
