% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_airflow.R
\name{prepare_framework}
\alias{prepare_framework}
\title{Prepare S3 operations and environment variables related to framework.}
\usage{
prepare_framework(estimator, s3_operations)
}
\arguments{
\item{estimator}{(sagemaker.estimator.Estimator): The framework estimator to
get information from and update.}

\item{s3_operations}{(list): The dict to specify s3 operations (upload
`source_dir`).}
}
\description{
S3 operations specify where to upload `source_dir`.
}
