% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_airflow.R
\name{prepare_framework_container_def}
\alias{prepare_framework_container_def}
\title{This prepares the framework model container information and specifies related S3 operations.}
\usage{
prepare_framework_container_def(model, instance_type, s3_operations)
}
\arguments{
\item{model}{(sagemaker.model.FrameworkModel): The framework model}

\item{instance_type}{(str): The EC2 instance type to deploy this Model to. For
example, 'ml.p2.xlarge'.}

\item{s3_operations}{(dict): The dict to specify S3 operations (upload
`source_dir` ).}
}
\value{
dict: The container information of this framework model.
}
\description{
Prepare the framework model container information. Specify related S3
             operations for Airflow to perform. (Upload `source_dir` )
}
