% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_conditions.R
\name{primitive_or_expr}
\alias{primitive_or_expr}
\title{Provide the expression of the value or return value if it is a primitive.}
\usage{
primitive_or_expr(value)
}
\arguments{
\item{value}{(Union[ConditionValueType, PrimitiveType]): The value to evaluate.}
}
\value{
Either the expression of the value or the primitive value.
}
\description{
Provide the expression of the value or return value if it is a primitive.
}
\keyword{internal}
