% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_airflow.R
\name{processing_config}
\alias{processing_config}
\title{Export Airflow processing config from a SageMaker processor}
\usage{
processing_config(
  processor,
  inputs = NULL,
  outputs = NULL,
  job_name = NULL,
  experiment_config = NULL,
  container_arguments = NULL,
  container_entrypoint = NULL,
  kms_key_id = NULL
)
}
\arguments{
\item{processor}{(sagemaker.processor.Processor): The SageMaker
processor to export Airflow config from.}

\item{inputs}{(list[:class:`~sagemaker.processing.ProcessingInput`]): Input files for
the processing job. These must be provided as
:class:`~sagemaker.processing.ProcessingInput` objects (default: None).}

\item{outputs}{(list[:class:`~sagemaker.processing.ProcessingOutput`]): Outputs for
the processing job. These can be specified as either path strings or
:class:`~sagemaker.processing.ProcessingOutput` objects (default: None).}

\item{job_name}{(str): Processing job name. If not specified, the processor generates
a default job name, based on the base job name and current timestamp.}

\item{experiment_config}{(dict[str, str]): Experiment management configuration.
Dictionary contains three optional keys:
'ExperimentName', 'TrialName', and 'TrialComponentDisplayName'.}

\item{container_arguments}{([str]): The arguments for a container used to run a processing job.}

\item{container_entrypoint}{([str]): The entrypoint for a container used to run a processing job.}

\item{kms_key_id}{(str): The AWS Key Management Service (AWS KMS) key that Amazon SageMaker
uses to encrypt the processing job output. KmsKeyId can be an ID of a KMS key,
ARN of a KMS key, alias of a KMS key, or alias of a KMS key.
The KmsKeyId is applied to all outputs.}
}
\value{
dict: Processing config that can be directly used by
           SageMakerProcessingOperator in Airflow.
}
\description{
Export Airflow processing config from a SageMaker processor
}
