% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_properties.R
\name{PropertiesList}
\alias{PropertiesList}
\title{Workflow PropertiesList class}
\description{
PropertiesList for use in workflow expressions
}
\section{Super classes}{
\code{\link[R6sagemaker.workflow:PropertiesMeta]{R6sagemaker.workflow::PropertiesMeta}} -> \code{\link[R6sagemaker.workflow:Properties]{R6sagemaker.workflow::Properties}} -> \code{PropertiesList}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PropertiesList$new()}}
\item \href{#method-get_item}{\code{PropertiesList$get_item()}}
\item \href{#method-clone}{\code{PropertiesList$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.workflow" data-topic="PropertiesMeta" data-id="format">}\href{../../R6sagemaker.workflow/html/PropertiesMeta.html#method-format}{\code{R6sagemaker.workflow::PropertiesMeta$format()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a PropertiesList instance representing the given shape.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PropertiesList$new(path, shape_name = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{(str): The parent path of the PropertiesList instance.}

\item{\code{shape_name}}{(str): The botocore sagemaker service model shape name.}

\item{\code{root_shape_name}}{(str): The botocore sagemaker service model shape name.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_item"></a>}}
\if{latex}{\out{\hypertarget{method-get_item}{}}}
\subsection{Method \code{get_item()}}{
Populate the indexing item with a Property, for both lists and dictionaries.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PropertiesList$get_item()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{item}}{(Union[int, str]): The index of the item in sequence.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PropertiesList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
