% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_properties.R
\name{PropertyFile}
\alias{PropertyFile}
\title{PropertyFile Class}
\description{
Provides a property file struct.
}
\section{Super class}{
\code{\link[sagemaker.workflow:Expression]{sagemaker.workflow::Expression}} -> \code{PropertyFile}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{The name of the property file for reference with `JsonGet` functions.}

\item{\code{output_name}}{The name of the processing job output channel.}

\item{\code{path}}{The path to the file at the output channel location.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{expr}}{Get the expression structure for workflow service calls.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PropertyFile$new()}}
\item \href{#method-clone}{\code{PropertyFile$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="sagemaker.workflow" data-topic="Expression" data-id="format">}\href{../../sagemaker.workflow/html/Expression.html#method-format}{\code{sagemaker.workflow::Expression$format()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initializing PropertyFile Class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PropertyFile$new(name, output_name, path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{(str): The name of the property file for reference with `JsonGet` functions.}

\item{\code{output_name}}{(str): The name of the processing job output channel.}

\item{\code{path}}{(str): The path to the file at the output channel location.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PropertyFile$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
