% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_quality_check_step.R
\name{QualityCheckConfig}
\alias{QualityCheckConfig}
\title{QualityCheckConfig class}
\description{
Quality Check Config.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{baseline_dataset}}{str or PipelineNonPrimitiveInputTypes): The path to the
baseline_dataset file. This can be a local path or an S3 uri string}

\item{\code{dataset_format}}{(dict): The format of the baseline_dataset.}

\item{\code{output_s3_uri}}{(str or PipelineNonPrimitiveInputTypes): Desired S3 destination of
the constraint_violations and statistics json files (default: None).
If not specified an auto generated path will be used:
"s3://<default_session_bucket>/model-monitor/baselining/<job_name>/results"}

\item{\code{post_analytics_processor_script}}{(str): The path to the record post-analytics
processor script (default: None). This can be a local path or an S3 uri string
but CANNOT be any of PipelineNonPrimitiveInputTypes.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-clone}{\code{QualityCheckConfig$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QualityCheckConfig$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
