% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_step_collections.R
\name{RegisterModel}
\alias{RegisterModel}
\title{Workflow RegisterModel class}
\description{
Register Model step collection for workflow.
}
\section{Super class}{
\code{\link[R6sagemaker.workflow:StepCollection]{R6sagemaker.workflow::StepCollection}} -> \code{RegisterModel}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{RegisterModel$new()}}
\item \href{#method-clone}{\code{RegisterModel$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.workflow" data-topic="StepCollection" data-id="format">}\href{../../R6sagemaker.workflow/html/StepCollection.html#method-format}{\code{R6sagemaker.workflow::StepCollection$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.workflow" data-topic="StepCollection" data-id="request_list">}\href{../../R6sagemaker.workflow/html/StepCollection.html#method-request_list}{\code{R6sagemaker.workflow::StepCollection$request_list()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Construct steps `_RepackModelStep` and `_RegisterModelStep` based on the estimator.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RegisterModel$new(
  name,
  content_types,
  response_types,
  inference_instances,
  transform_instances,
  estimator = NULL,
  model_data = NULL,
  depends_on = NULL,
  model_package_group_name = NULL,
  model_metrics = NULL,
  approval_status = NULL,
  image_uri = NULL,
  compile_model_family = NULL,
  description = NULL,
  tags = NULL,
  model = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{(str): The name of the training step.}

\item{\code{content_types}}{(list): The supported MIME types for the input data (default: None).}

\item{\code{response_types}}{(list): The supported MIME types for the output data (default: None).}

\item{\code{inference_instances}}{(list): A list of the instance types that are used to
generate inferences in real-time (default: None).}

\item{\code{transform_instances}}{(list): A list of the instance types on which a transformation
job can be run or on which an endpoint can be deployed (default: None).}

\item{\code{estimator}}{: The estimator instance.}

\item{\code{model_data}}{: The S3 uri to the model data from training.}

\item{\code{depends_on}}{(List[str]): The list of step names the first step in the collection
depends on}

\item{\code{model_package_group_name}}{(str): The Model Package Group name, exclusive to
`model_package_name`, using `model_package_group_name` makes the Model Package
versioned (default: None).}

\item{\code{model_metrics}}{(ModelMetrics): ModelMetrics object (default: None).}

\item{\code{approval_status}}{(str): Model Approval Status, values can be "Approved", "Rejected",
or "PendingManualApproval" (default: "PendingManualApproval").}

\item{\code{image_uri}}{(str): The container image uri for Model Package, if not specified,
Estimator's training container image is used (default: None).}

\item{\code{compile_model_family}}{(str): The instance family for the compiled model. If
specified, a compiled model is used (default: None).}

\item{\code{description}}{(str): Model Package description (default: None).}

\item{\code{tags}}{(List[dict[str, str]]): The list of tags to attach to the model package group. Note
that tags will only be applied to newly created model package groups; if the
name of an existing group is passed to "model_package_group_name",
tags will not be applied.}

\item{\code{model}}{(object or Model): A PipelineModel object that comprises a list of models
which gets executed as a serial inference pipeline or a Model object.}

\item{\code{...}}{: additional arguments to `create_model`.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RegisterModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
