% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_retry.R
\docType{data}
\name{SageMakerJobExceptionTypeEnum}
\alias{SageMakerJobExceptionTypeEnum}
\title{SageMaker Job ExceptionType enum.}
\format{
An object of class \code{Enum} (inherits from \code{environment}) of length 3.
}
\usage{
SageMakerJobExceptionTypeEnum
}
\description{
SageMaker Job ExceptionType enum.
}
\keyword{datasets}
