% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_steps.R
\name{Step}
\alias{Step}
\title{Workflow Step class}
\description{
Pipeline step for workflow.
}
\section{Super class}{
\code{\link[R6sagemaker.workflow:Entity]{R6sagemaker.workflow::Entity}} -> \code{Step}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{The name of the step.}

\item{\code{step_type}}{The type of the step.}

\item{\code{depends_on}}{The list of step names the current step depends on}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{arguments}}{The arguments to the particular step service call.}

\item{\code{properties}}{The properties of the particular step.}

\item{\code{ref}}{Gets a reference dict for steps}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Step$new()}}
\item \href{#method-to_request}{\code{Step$to_request()}}
\item \href{#method-add_depends_on}{\code{Step$add_depends_on()}}
\item \href{#method-format}{\code{Step$format()}}
\item \href{#method-clone}{\code{Step$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize Workflow Step
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Step$new(name, step_type = enum_items(StepTypeEnum), depends_on = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{(str): The name of the step.}

\item{\code{step_type}}{(StepTypeEnum): The type of the step.}

\item{\code{depends_on}}{(List[str]): The list of step names the current step depends on}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_request"></a>}}
\if{latex}{\out{\hypertarget{method-to_request}{}}}
\subsection{Method \code{to_request()}}{
Gets the request structure for workflow service calls.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Step$to_request()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_depends_on"></a>}}
\if{latex}{\out{\hypertarget{method-add_depends_on}{}}}
\subsection{Method \code{add_depends_on()}}{
Add step names to the current step depends on list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Step$add_depends_on(step_names)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{step_names}}{(list): placeholder}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-format"></a>}}
\if{latex}{\out{\hypertarget{method-format}{}}}
\subsection{Method \code{format()}}{
formats class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Step$format()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Step$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
