% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_step_collections.R
\name{StepCollection}
\alias{StepCollection}
\title{Workflow StepCollection class}
\description{
A wrapper of pipeline steps for workflow.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{steps}}{A list of steps.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{StepCollection$new()}}
\item \href{#method-request_list}{\code{StepCollection$request_list()}}
\item \href{#method-format}{\code{StepCollection$format()}}
\item \href{#method-clone}{\code{StepCollection$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize StepCollection class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StepCollection$new(steps)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{steps}}{(List[Step]): A list of steps.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-request_list"></a>}}
\if{latex}{\out{\hypertarget{method-request_list}{}}}
\subsection{Method \code{request_list()}}{
Get the request structure for workflow service calls.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StepCollection$request_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-format"></a>}}
\if{latex}{\out{\hypertarget{method-format}{}}}
\subsection{Method \code{format()}}{
format class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StepCollection$format()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StepCollection$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
