% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_retry.R
\docType{data}
\name{StepExceptionTypeEnum}
\alias{StepExceptionTypeEnum}
\title{Step ExceptionType enum.}
\format{
An object of class \code{Enum} (inherits from \code{environment}) of length 2.
}
\usage{
StepExceptionTypeEnum
}
\description{
Step ExceptionType enum.
}
\keyword{datasets}
