% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_retry.R
\name{StepRetryPolicy}
\alias{StepRetryPolicy}
\title{StepRetryPolicy class}
\description{
RetryPolicy for a retryable step. The pipeline service will retry
}
\section{Super classes}{
\code{\link[sagemaker.workflow:Entity]{sagemaker.workflow::Entity}} -> \code{\link[sagemaker.workflow:RetryPolicy]{sagemaker.workflow::RetryPolicy}} -> \code{StepRetryPolicy}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{exception_types}}{(List[StepExceptionTypeEnum]): the exception types to match for this policy}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{StepRetryPolicy$new()}}
\item \href{#method-to_request}{\code{StepRetryPolicy$to_request()}}
\item \href{#method-clone}{\code{StepRetryPolicy$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="sagemaker.workflow" data-topic="RetryPolicy" data-id="format">}\href{../../sagemaker.workflow/html/RetryPolicy.html#method-format}{\code{sagemaker.workflow::RetryPolicy$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sagemaker.workflow" data-topic="RetryPolicy" data-id="validate_backoff_rate">}\href{../../sagemaker.workflow/html/RetryPolicy.html#method-validate_backoff_rate}{\code{sagemaker.workflow::RetryPolicy$validate_backoff_rate()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sagemaker.workflow" data-topic="RetryPolicy" data-id="validate_expire_after_mins">}\href{../../sagemaker.workflow/html/RetryPolicy.html#method-validate_expire_after_mins}{\code{sagemaker.workflow::RetryPolicy$validate_expire_after_mins()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sagemaker.workflow" data-topic="RetryPolicy" data-id="validate_interval_seconds">}\href{../../sagemaker.workflow/html/RetryPolicy.html#method-validate_interval_seconds}{\code{sagemaker.workflow::RetryPolicy$validate_interval_seconds()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sagemaker.workflow" data-topic="RetryPolicy" data-id="validate_max_attempts">}\href{../../sagemaker.workflow/html/RetryPolicy.html#method-validate_max_attempts}{\code{sagemaker.workflow::RetryPolicy$validate_max_attempts()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize StepRetryPolicy class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StepRetryPolicy$new(
  exception_types,
  backoff_rate = 2,
  interval_seconds = 1,
  max_attempts = NULL,
  expire_after_mins = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{exception_types}}{(List[StepExceptionTypeEnum]): the exception types
to match for this policy}

\item{\code{backoff_rate}}{(float): The multiplier by which the retry interval increases
during each attempt (default: 2.0)}

\item{\code{interval_seconds}}{(int): An integer that represents the number of seconds before the
first retry attempt (default: 1)}

\item{\code{max_attempts}}{(int): A positive integer that represents the maximum
number of retry attempts. (default: None)}

\item{\code{expire_after_mins}}{(int): A positive integer that represents the maximum minute
to expire any further retry attempt (default: None)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_request"></a>}}
\if{latex}{\out{\hypertarget{method-to_request}{}}}
\subsection{Method \code{to_request()}}{
Gets the request structure for retry policy.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StepRetryPolicy$to_request()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StepRetryPolicy$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
