% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_steps.R
\docType{data}
\name{StepTypeEnum}
\alias{StepTypeEnum}
\title{Workflow StepTypeEnum class}
\format{
An object of class \code{Enum} (inherits from \code{environment}) of length 8.
}
\usage{
StepTypeEnum
}
\description{
Enum of step types.
}
\keyword{datasets}
