% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_airflow.R
\name{update_estimator_from_task}
\alias{update_estimator_from_task}
\title{Update training job of the estimator from a task in the DAG}
\usage{
update_estimator_from_task(estimator, task_id, task_type)
}
\arguments{
\item{estimator}{(sagemaker.estimator.EstimatorBase): The estimator to update}

\item{task_id}{(str): The task id of any
airflow.contrib.operators.SageMakerTrainingOperator or
airflow.contrib.operators.SageMakerTuningOperator that generates
training jobs in the DAG.}

\item{task_type}{(str): Whether the task is from SageMakerTrainingOperator or
SageMakerTuningOperator. Values can be 'training', 'tuning' or None
(which means training job is not from any task).}
}
\description{
Update training job of the estimator from a task in the DAG
}
