% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_airflow.R
\name{update_submit_s3_uri}
\alias{update_submit_s3_uri}
\title{Updated the S3 URI of the framework source directory in given estimator.}
\usage{
update_submit_s3_uri(estimator, job_name)
}
\arguments{
\item{estimator}{(sagemaker.estimator.Framework): The Framework estimator to
update.}

\item{job_name}{(str): The new job name included in the submit S3 URI}
}
\value{
str: The updated S3 URI of framework source directory
}
\description{
Updated the S3 URI of the framework source directory in given estimator.
}
