% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_step_collections.R
\name{EstimatorTransformer}
\alias{EstimatorTransformer}
\title{Workflow EstimatorTransformer class}
\description{
Creates a Transformer step collection for workflow.
}
\section{Super class}{
\code{\link[sagemaker.workflow:StepCollection]{sagemaker.workflow::StepCollection}} -> \code{EstimatorTransformer}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{EstimatorTransformer$new()}}
\item \href{#method-clone}{\code{EstimatorTransformer$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="sagemaker.workflow" data-topic="StepCollection" data-id="format">}\href{../../sagemaker.workflow/html/StepCollection.html#method-format}{\code{sagemaker.workflow::StepCollection$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sagemaker.workflow" data-topic="StepCollection" data-id="request_list">}\href{../../sagemaker.workflow/html/StepCollection.html#method-request_list}{\code{sagemaker.workflow::StepCollection$request_list()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Construct steps required for a Transformer step collection:
             An estimator-centric step collection. It models what happens in workflows
             when invoking the `transform()` method on an estimator instance:
             First, if custom model artifacts are required, a `_RepackModelStep` is included.
             Second, a `CreateModelStep` with the model data passed in from
             a training step or other training job output. Finally, a `TransformerStep`.
             If repacking the model artifacts is not necessary, only the CreateModelStep
             and TransformerStep are in the step collection.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EstimatorTransformer$new(
  name,
  estimator,
  model_data,
  model_inputs,
  instance_count,
  instance_type,
  transform_inputs,
  description = NULL,
  display_name = NULL,
  image_uri = NULL,
  predictor_cls = NULL,
  env = NULL,
  strategy = NULL,
  assemble_with = NULL,
  output_path = NULL,
  output_kms_key = NULL,
  accept = NULL,
  max_concurrent_transforms = NULL,
  max_payload = NULL,
  tags = NULL,
  volume_kms_key = NULL,
  depends_on = NULL,
  repack_model_step_retry_policies = NULL,
  model_step_retry_policies = NULL,
  transform_step_retry_policies = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{(str): The name of the Transform Step.}

\item{\code{estimator}}{: The estimator instance.}

\item{\code{model_data}}{(str): The S3 location of a SageMaker model data
``.tar.gz`` file (default: None).}

\item{\code{model_inputs}}{(CreateModelInput): A `sagemaker.inputs.CreateModelInput` instance.
Defaults to `None`.}

\item{\code{instance_count}}{(int): The number of EC2 instances to use.}

\item{\code{instance_type}}{(str): The type of EC2 instance to use.}

\item{\code{transform_inputs}}{(TransformInput): A `sagemaker.inputs.TransformInput` instance.}

\item{\code{description}}{(str): The description of the step.}

\item{\code{display_name}}{(str): The display name of the step.}

\item{\code{image_uri}}{(str): A Docker image URI.}

\item{\code{predictor_cls}}{(callable[string, :Session]): A
function to call to create a predictor (default: None). If not
None, ``deploy`` will return the result of invoking this
function on the created endpoint name.}

\item{\code{env}}{(dict): The Environment variables to be set for use during the
transform job (default: None).}

\item{\code{strategy}}{(str): The strategy used to decide how to batch records in
a single request (default: None). Valid values: 'MultiRecord'
and 'SingleRecord'.}

\item{\code{assemble_with}}{(str): How the output is assembled (default: None).
Valid values: 'Line' or 'None'.}

\item{\code{output_path}}{(str): The S3 location for saving the transform result. If
not specified, results are stored to a default bucket.}

\item{\code{output_kms_key}}{(str): Optional. A KMS key ID for encrypting the
transform output (default: None).}

\item{\code{accept}}{(str): The accept header passed by the client to
the inference endpoint. If it is supported by the endpoint,
it will be the format of the batch transform output.}

\item{\code{max_concurrent_transforms}}{(int): The maximum number of HTTP requests
to be made to each individual transform container at one time.}

\item{\code{max_payload}}{(int): Maximum size of the payload in a single HTTP}

\item{\code{tags}}{(list[dict]): List of tags for labeling a training job. For
more, see \url{https://docs.aws.amazon.com/sagemaker/latest/dg/API_Tag.html}.}

\item{\code{volume_kms_key}}{(str): Optional. KMS key ID for encrypting the volume
attached to the ML compute instance (default: None).}

\item{\code{depends_on}}{(List[str] or List[Step]): The list of step names or step instances
the first step in the collection depends on}

\item{\code{repack_model_step_retry_policies}}{(List[RetryPolicy]): The list of retry policies
for the repack model step}

\item{\code{model_step_retry_policies}}{(List[RetryPolicy]): The list of retry policies for
model step}

\item{\code{transform_step_retry_policies}}{(List[RetryPolicy]): The list of retry policies for
transform step}

\item{\code{...}}{: pass onto model class.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EstimatorTransformer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
