% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_execution_variables.R
\docType{data}
\name{ExecutionVariables}
\alias{ExecutionVariables}
\title{Enum-like class for all ExecutionVariable instances.}
\format{
An object of class \code{Enum} (inherits from \code{environment}) of length 6.
}
\usage{
ExecutionVariables
}
\description{
Considerations to move these as module-level constants should be made.
}
\keyword{datasets}
