% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_condition_step.R
\name{JsonGet}
\alias{JsonGet}
\title{Workflow JsonGet class}
\description{
Get JSON properties from PropertyFiles.
}
\section{Super class}{
\code{\link[sagemaker.workflow:Expression]{sagemaker.workflow::Expression}} -> \code{JsonGet}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{step_name}}{The step from which to get the property file.}

\item{\code{property_file}}{Either a PropertyFile instance or the name of a property file.}

\item{\code{json_path}}{The JSON path expression to the requested value.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{expr}}{The expression dict for a `JsonGet` function.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{JsonGet$new()}}
\item \href{#method-clone}{\code{JsonGet$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="sagemaker.workflow" data-topic="Expression" data-id="format">}\href{../../sagemaker.workflow/html/Expression.html#method-format}{\code{sagemaker.workflow::Expression$format()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize JsonGet class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JsonGet$new(step_name, property_file, json_path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{step_name}}{(Step): The step from which to get the property file.}

\item{\code{property_file}}{(Union[PropertyFile, str]): Either a PropertyFile instance
or the name of a property file.}

\item{\code{json_path}}{(str): The JSON path expression to the requested value.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JsonGet$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
