% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_steps.R
\name{ProcessingStep}
\alias{ProcessingStep}
\title{Workflow ProcessingStep Class}
\description{
Processing step for workflow.
}
\section{Super classes}{
\code{\link[sagemaker.workflow:Entity]{sagemaker.workflow::Entity}} -> \code{\link[sagemaker.workflow:Step]{sagemaker.workflow::Step}} -> \code{\link[sagemaker.workflow:ConfigurableRetryStep]{sagemaker.workflow::ConfigurableRetryStep}} -> \code{ProcessingStep}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{arguments}}{The arguments dict that is used to call `create_processing_job`.
NOTE: The CreateProcessingJob request is not quite the args list that workflow needs.
ProcessingJobName and ExperimentConfig cannot be included in the arguments.}

\item{\code{properties}}{A Properties object representing the DescribeProcessingJobResponse data model.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ProcessingStep$new()}}
\item \href{#method-to_request}{\code{ProcessingStep$to_request()}}
\item \href{#method-clone}{\code{ProcessingStep$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="sagemaker.workflow" data-topic="Step" data-id="add_depends_on">}\href{../../sagemaker.workflow/html/Step.html#method-add_depends_on}{\code{sagemaker.workflow::Step$add_depends_on()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sagemaker.workflow" data-topic="Step" data-id="format">}\href{../../sagemaker.workflow/html/Step.html#method-format}{\code{sagemaker.workflow::Step$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sagemaker.workflow" data-topic="ConfigurableRetryStep" data-id="add_retry_policy">}\href{../../sagemaker.workflow/html/ConfigurableRetryStep.html#method-add_retry_policy}{\code{sagemaker.workflow::ConfigurableRetryStep$add_retry_policy()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Construct a ProcessingStep, given a `Processor` instance.
             In addition to the processor instance, the other arguments are those that are supplied to
             the `process` method of the `sagemaker.processing.Processor`.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProcessingStep$new(
  name,
  processor,
  display_name = NULL,
  description = NULL,
  inputs = NULL,
  outputs = NULL,
  job_arguments = NULL,
  code = NULL,
  property_files = NULL,
  cache_config = NULL,
  depends_on = NULL,
  retry_policies = NULL,
  kms_key = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{(str): The name of the processing step.}

\item{\code{processor}}{(Processor): A `sagemaker.processing.Processor` instance.}

\item{\code{display_name}}{(str): The display name of the processing step.}

\item{\code{description}}{(str): The description of the processing step.}

\item{\code{inputs}}{(List[ProcessingInput]): A list of `sagemaker.processing.ProcessorInput`
instances. Defaults to `None`.}

\item{\code{outputs}}{(List[ProcessingOutput]): A list of `sagemaker.processing.ProcessorOutput`
instances. Defaults to `None`.}

\item{\code{job_arguments}}{(List[str]): A list of strings to be passed into the processing job.
Defaults to `None`.}

\item{\code{code}}{(str): This can be an S3 URI or a local path to a file with the framework
script to run. Defaults to `None`.}

\item{\code{property_files}}{(List[PropertyFile]): A list of property files that workflow looks
for and resolves from the configured processing output list.}

\item{\code{cache_config}}{(CacheConfig):  A `sagemaker.workflow.steps.CacheConfig` instance.}

\item{\code{depends_on}}{(List[str] or List[Step]): A list of step names or step instance
this `sagemaker.workflow.steps.ProcessingStep` depends on}

\item{\code{retry_policies}}{(List[RetryPolicy]):  A list of retry policy}

\item{\code{kms_key}}{(str): The ARN of the KMS key that is used to encrypt the
user code file. Defaults to `None`.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_request"></a>}}
\if{latex}{\out{\hypertarget{method-to_request}{}}}
\subsection{Method \code{to_request()}}{
Get the request structure for workflow service calls.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProcessingStep$to_request()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProcessingStep$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
